Fix logic error in print_buffer function
The conflicting pipeline was not being displayed properly b/c the inst. in buffer[0] was not being replaced by the inst. that would actually was going to end up in the pipeline causing the instruction in buffer[0] to appear is if it was in the pipeline twice when that is not actually the case